releasing 2.1.1 patch
This release is a backport of https://github.com/ef4/fast-sourcemap-concat/pull/72 on top of 2.10.

master has other changes that have sat for a long time and they really should go out in a major, dropping ancient node versions and some deprecated dependencies. But I want to do this patch release too so that people will get the fix automatically, even on older versions.